Finish basis function.
This branch does not change pil to pillow, for we still need matplotlib
to plot data. So Pyhton 2.7 is necessary.

The .py could list all the .png files in a designated dir_path, read the
.png and output a .txt, seprately.
The .txt includes frequency as x-axis and dB(uV/m) as y-axis. The range
of frequency is 30MHz to 300MHz.